console: Streamline application add form
